requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
pillow
opencv-python
matplotlib
seaborn
tqdm
scipy
